tidy up mg25 tests
